servoshell: Move screenshot stability tracking to shared base
